Move Multi-SEM clients and tests to corresponding packages
